refactor: rename fab to fabTabNavigation
