--- src/components/Fab/index.ts
@@ -1,1 +1,0 @@
-export * from './Fab';
