+++ src/components/FabTabNavigation/FabTabNavigation.spec.tsx
@@ -1,11 +1,11 @@
 import React from 'react';
 import { render } from '@testing-library/react-native';
-import { Fab } from './Fab';
+import { FabTabNavigation } from './FabTabNavigation';
 import { wrapper } from '@app/utils/testUtils';
 
 describe('<Fab />', () => {
   it('Should render correctly', () => {
-    const tree = render(<Fab />, { wrapper }).toJSON();
+    const tree = render(<FabTabNavigation />, { wrapper }).toJSON();
 
     expect(tree).toMatchSnapshot();
   });
