+++ src/navigation/tabNavigation/TabNavigation.tsx
@@ -6,7 +6,7 @@ import { Feather } from '@expo/vector-icons';
 
 import { Menu } from '@app/screens';
 import { colors, fonts } from '@app/theme';
-import { Fab } from '@app/components';
+import { FabTabNavigation } from '@app/components';
 import { ShoppingListsStackNavigator } from '../stackNavigation/shoppingList';
 
 export function TabNavigator() {
@@ -39,10 +39,10 @@ export function TabNavigator() {
           headerTitleAlign: 'left',
         };
       }}
-      initialRouteName={'Home'}
+      initialRouteName={'HomeTab'}
     >
       <Tab.Screen
-        name="Home"
+        name="HomeTab"
         component={ShoppingListsStackNavigator}
         options={{
           tabBarLabel: 'Home',
@@ -56,7 +56,7 @@ export function TabNavigator() {
         options={{
           tabBarButton: () => (
             <View style={{ marginTop: -16 }}>
-              <Fab />
+              <FabTabNavigation />
             </View>
           ),
         }}
